[Silver II] Title: 랜선 자르기, Time: 8 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -16,7 +16,7 @@ int main()
 	for (int i = 0; i < k; i++) { cin >> lan[i]; }
 	sort(lan.begin(), lan.end());
 
-	long long low, high, mid, sum;
+	long low, high, mid, sum;
 
 	low = 1;
 	high = lan[k - 1];
